requirements.txt
streamlit
pypdf
faiss-cpu
numpy
sentence-transformers
google-generativeai>=0.7.2
torch
python-dotenv